test_bot.py
#!/usr/bin/env python3
"""
Test script for Video Downloader Bot
Tests all major components without requiring actual Telegram connection
"""

import sys
import os

print("🧪 Testing Video Downloader Bot Components...\n")

# Test 1: Environment Variables
print("=" * 60)
print("TEST 1: Configuration Loading")
print("=" * 60)

try:
    from dotenv import load_dotenv
    load_dotenv()
    print("✅ dotenv loaded successfully")
except ImportError:
    print("⚠️  python-dotenv not installed (optional)")

try:
    from config import BotConfig
    
    # Test with environment variable
    if os.getenv("BOT_TOKEN"):
        config = BotConfig.from_env()
        print("✅ Configuration loaded from environment")
        print(f"   - Max file size: {config.max_file_size / (1024*1024):.0f}MB")
        print(f"   - Download timeout: {config.download_timeout}s")
        print(f"   - Rate limit: {config.rate_limit_requests} requests per {config.rate_limit_period}s")
        print(f"   - Log level: {config.log_level}")
    else:
        print("⚠️  BOT_TOKEN not set - will fail in production")
        print("   Set it with: export BOT_TOKEN='your_token_here'")
except Exception as e:
    print(f"❌ Configuration test failed: {e}")
    sys.exit(1)

# Test 2: Custom Exceptions
print("\n" + "=" * 60)
print("TEST 2: Custom Exceptions")
print("=" * 60)

try:
    from exceptions import (
        BotException, DownloadError, ValidationError, 
        RateLimitError, FileSizeError, ConfigurationError
    )
    
    # Test exception hierarchy
    test_exceptions = [
        DownloadError("Test download error"),
        ValidationError("Test validation error"),
        RateLimitError("Test rate limit error"),
        FileSizeError("Test file size error"),
    ]
    
    for exc in test_exceptions:
        assert isinstance(exc, BotException), f"{type(exc).__name__} not a BotException"
    
    print("✅ All custom exceptions working")
    print(f"   - Defined exceptions: {len(test_exceptions)}")
    
except Exception as e:
    print(f"❌ Exception test failed: {e}")
    sys.exit(1)

# Test 3: URL Validation
print("\n" + "=" * 60)
print("TEST 3: URL Validation")
print("=" * 60)

try:
    from validators import URLValidator
    
    test_urls = [
        ("https://youtube.com/watch?v=dQw4w9WgXcQ", True, "youtube"),
        ("https://youtu.be/dQw4w9WgXcQ", True, "youtube"),
        ("https://tiktok.com/@user/video/123456", True, "tiktok"),
        ("https://vm.tiktok.com/ZMhKFqxyz/", True, "tiktok"),
        ("https://instagram.com/p/ABC123/", True, "instagram"),
        ("https://instagram.com/reel/XYZ789/", True, "instagram"),
        ("https://invalid-site.com/video", False, None),
        ("not_a_url", False, None),
    ]
    
    passed = 0
    failed = 0
    
    for url, should_be_valid, expected_platform in test_urls:
        is_valid, platform, error = URLValidator.validate(url)
        
        if is_valid == should_be_valid and (not should_be_valid or platform == expected_platform):
            passed += 1
            status = "✅"
        else:
            failed += 1
            status = "❌"
        
        print(f"   {status} {url[:50]:50} -> {platform or 'invalid'}")
    
    print(f"\n✅ URL Validation: {passed}/{len(test_urls)} tests passed")
    
    if failed > 0:
        print(f"⚠️  {failed} validation tests failed")
    
except Exception as e:
    print(f"❌ URL validation test failed: {e}")
    sys.exit(1)

# Test 4: Input Sanitization
print("\n" + "=" * 60)
print("TEST 4: Input Sanitization")
print("=" * 60)

try:
    from validators import InputSanitizer
    
    test_inputs = [
        ("normal_filename.mp4", "normal_filename.mp4"),
        ("../../../etc/passwd", ".._.._.._.._etc_passwd"),
        ("file<>:|?*.txt", "file_______.txt"),
        ("very" * 100 + ".mp4", True),  # Should be truncated
        ("   spaces   .mp4", "spaces.mp4"),
    ]
    
    passed = 0
    for original, expected in test_inputs:
        sanitized = InputSanitizer.sanitize_filename(original)
        
        # Check if it's safe
        if "../" not in sanitized and "\\" not in sanitized:
            passed += 1
            print(f"   ✅ {original[:30]:30} -> {sanitized[:30]}")
        else:
            print(f"   ❌ {original[:30]:30} -> {sanitized[:30]} (UNSAFE)")
    
    print(f"\n✅ Input Sanitization: {passed}/{len(test_inputs)} tests passed")
    
except Exception as e:
    print(f"❌ Sanitization test failed: {e}")
    sys.exit(1)

# Test 5: Rate Limiter
print("\n" + "=" * 60)
print("TEST 5: Rate Limiter")
print("=" * 60)

try:
    from rate_limiter import RateLimiter
    
    limiter = RateLimiter(user_capacity=3, user_refill_rate=1.0)
    
    user_id = 12345
    
    # Should allow first 3 requests
    results = []
    for i in range(5):
        allowed, wait_time = limiter.check_limit(user_id)
        results.append((i+1, allowed, wait_time))
    
    print("   Request Results:")
    for req_num, allowed, wait_time in results:
        status = "✅ Allowed" if allowed else f"❌ Blocked (wait {wait_time:.1f}s)"
        print(f"   Request {req_num}: {status}")
    
    # Check that first 3 passed and rest blocked
    allowed_count = sum(1 for _, allowed, _ in results if allowed)
    blocked_count = sum(1 for _, allowed, _ in results if not allowed)
    
    if allowed_count == 3 and blocked_count == 2:
        print("\n✅ Rate limiter working correctly")
    else:
        print(f"\n⚠️  Rate limiter: {allowed_count} allowed, {blocked_count} blocked")
    
except Exception as e:
    print(f"❌ Rate limiter test failed: {e}")
    sys.exit(1)

# Test 6: Resource Manager
print("\n" + "=" * 60)
print("TEST 6: Resource Manager")
print("=" * 60)

try:
    from resource_manager import ResourceManager
    import asyncio
    
    async def test_resource_manager():
        manager = ResourceManager(max_concurrent_downloads=2, max_downloads_per_user=1)
        
        user1 = 12345
        user2 = 67890
        
        # Test concurrent limit
        try:
            async with manager.download_slot(user1):
                print("   ✅ User 1 acquired slot 1")
                
                async with manager.download_slot(user2):
                    print("   ✅ User 2 acquired slot 2")
                    
                    # This should fail (global limit)
                    try:
                        async with manager.download_slot(user1):
                            print("   ❌ Should not reach here (global limit)")
                    except Exception as e:
                        print(f"   ✅ Correctly blocked 3rd download: {type(e).__name__}")
        except Exception as e:
            print(f"   ⚠️  Resource manager error: {e}")
        
        # Check status
        status = await manager.get_status()
        print(f"\n   Status: {status['active_downloads']}/{status['max_downloads']} active")
        
        return True
    
    # Run async test
    result = asyncio.run(test_resource_manager())
    
    if result:
        print("\n✅ Resource manager working correctly")
    
except Exception as e:
    print(f"❌ Resource manager test failed: {e}")
    sys.exit(1)

# Test 7: Utilities
print("\n" + "=" * 60)
print("TEST 7: Utility Functions")
print("=" * 60)

try:
    from utils import format_bytes, format_duration, validate_quality
    
    # Test format_bytes
    test_bytes = [
        (0, "0 B"),
        (1024, "1.0 KB"),
        (1024 * 1024, "1.0 MB"),
        (1024 * 1024 * 1024, "1.0 GB"),
    ]
    
    print("   Format Bytes:")
    for bytes_val, expected in test_bytes:
        result = format_bytes(bytes_val)
        status = "✅" if expected in result else "⚠️"
        print(f"   {status} {bytes_val:15} -> {result}")
    
    # Test format_duration
    print("\n   Format Duration:")
    test_durations = [
        (30, "0:30"),
        (90, "1:30"),
        (3600, "1:00:00"),
    ]
    
    for seconds, expected in test_durations:
        result = format_duration(seconds)
        status = "✅" if expected in result else "⚠️"
        print(f"   {status} {seconds:6}s -> {result}")
    
    # Test validate_quality
    print("\n   Validate Quality:")
    for quality in ['360p', 'audio', 'invalid']:
        result = validate_quality(quality)
        status = "✅" if (result and quality != 'invalid') or (not result and quality == 'invalid') else "❌"
        print(f"   {status} '{quality}' -> {result}")
    
    print("\n✅ All utility functions working")
    
except Exception as e:
    print(f"❌ Utilities test failed: {e}")
    sys.exit(1)

# Test 8: Import bot.py
print("\n" + "=" * 60)
print("TEST 8: Bot Module Import")
print("=" * 60)

try:
    # Only test if BOT_TOKEN is set (won't crash without it now)
    if os.getenv("BOT_TOKEN"):
        print("   Attempting to import bot module...")
        import bot
        print("   ✅ Bot module imported successfully")
        print(f"   ✅ Config loaded: {bot.config.max_file_size / (1024*1024):.0f}MB limit")
        print(f"   ✅ Rate limiter initialized")
        print(f"   ✅ Resource manager initialized")
    else:
        print("   ⚠️  Skipping bot import (BOT_TOKEN not set)")
        print("   Set BOT_TOKEN to test full bot import")
    
except Exception as e:
    print(f"   ❌ Bot import failed: {e}")
    print(f"   This might be normal if Telegram libraries aren't fully set up")

# Final Summary
print("\n" + "=" * 60)
print("TEST SUMMARY")
print("=" * 60)

if os.getenv("BOT_TOKEN"):
    print("✅ All core components tested successfully!")
    print("\n🚀 Your bot is ready to run:")
    print("   python bot.py")
else:
    print("⚠️  Most components tested successfully!")
    print("\n⚠️  To complete testing, set BOT_TOKEN:")
    print("   export BOT_TOKEN='your_actual_token_here'")
    print("   python test_bot.py")

print("\n📚 Documentation:")
print("   - README.md - Full documentation")
print("   - QUICK_START.md - 5-minute setup guide")
print("   - IMPROVEMENTS.md - All improvements listed")

print("\n" + "=" * 60)
