requirements.txt
python-telegram-bot>=21.0
yt-dlp>=2024.8.6
ffmpeg-python==0.2.0
psutil>=5.9.0
requests>=2.31.0
python-dotenv>=1.0.0
